Made InfoBox slide in and out as need be.
It'll slide in when ExpeditionMode starts and slide out when it ends.

Also, Jumbo Mode for the infobox is no longer a thing.  The prefs have been
updated accordingly, though I'll work on clearing out the old one later.
The only places it's used are in old classes I don't need to maintain.