Fixed weird date issues with SelectAGraticuleMode.
If the user winds up with an invalid date in SelectAGraticuleMode, it'll
now just revert to the last known "good" date and graticule combination
when returning to ExpeditionMode.  Otherwise, ExpeditionMode would start
off in an invalid state, which is confusing.

Notably, ExpeditionMode CAN still start in an invalid state, but now it
should only be if the user starts the app fresh before the DJIA value for
the day is posted.  That's okay; I just didn't want the user to be able to
screw it up by accident.

Also, I should do the same thing in ExpeditionMode, as that also handles
date changes.