Ironed out more of DetailedInfoActivity.
I'm starting to wonder why I did the weird "pass in a GoogleApiClient and
let CentralMap deal with the connection" thing in CentralMapMode, because
it makes a lot more sense, in DetailedInfoFragment at least, to just make
a new client there and use it.

Regardless, DetailedInfoActivity is now far more fleshed out and stable.
It still needs a menu, but the main portion of it is good.  Next, doing the
same thing but with the Fragment embedded in the map screen.