Added Send to Maps to ExpeditionMode's menu.
On second thought, I should keep it there, just for consistency's sake.
And so that there's a way to send it to something else that understands the
intent.

Also, CentralMapMode's onCreateOptionsMenu now takes the context.  I found
it's possible for that to be called before the mode has the CentralMap
reference handy, which, of course, leads to our good friend, the
NullPointerException.