Created MapTypeDialogFragment.
This comes in when you pick the new menu option for the map type.  I
couldn't fit another set of icons anywhere on the map.

Note that this isn't styled like the usual Geohash Droid dialogs yet.
Not quite sure what AlertDialog.Builder is looking for to make sure the
list will work.