Updated the wiki image infobox.
Now the infobox painted on images will have the same icons as what's on the
main interface.  It also cleans up some of the positioning logic now that
it's years later and I know that Android draws text from the baseline-left
point, not top-left.

Also, wiki image quality has been bumped up to 90 instead of 85.  That
doesn't seem to be going over the 150kb limit on the wiki (last I heard; I
really should see if that's changed over the years), and it makes things
look a bit better with the scaled-down icons.