Initial commit of the wiki fragment layout.
Okay, I think that'll do.  This time around, there's only going to be one
wiki posting screen.  A checkbox will indicate if it's a picture post or
just plain text.  It'll go in the same spot as the detailed info fragment
on tablets, or it'll be a full activity for phones.

Because of that last part, the FrameLayout that holds the detailed info
fragment has been renamed to something hopefully more useful.