Made WikiImageUtils.ImageInfo more URI-friendly.
Since we're now dealing with the reasonable possibility that MediaStore
WON'T know what to do with the images we feed it, readImageInfo has been
updated to account for this.  Now it'll ALWAYS return a valid ImageInfo, at
least as far as the URI, location, and timestamp are concerned.

WikiService has also been updated to account for this.  And it works, too.
Woo!