Made it possible to get to WikiFragment.
This required more rearranging than I really thought should have been
necessary, but it appears to work.  The phone layout will launch a new
Activity for it and the tablet layout will re-use the same container that
DetailedInfoFragment uses.

There's also CentralMapExtraFragment as an abstract base class, an enum,
some changes to ExpeditionMode, more specifics with the menu, it's all just
a huge mess.

Note that WikiFragment doesn't really WORK yet, but you can GET to it, at
the very least.