More WikiFragment work.
WikiFragment will now properly onSaveInstanceState the picture URI, for
starters.  There's also going to be a header on top, showing the date and
the graticule (or globalhash).

Still need to make it DO something, of course.  And I'm not quite sure yet
how I'll provide a link to the expedition page itself...