Cleaned up some old classes and API v1 calls.
All the old Maps API v1 stuff should be cleaned out now.  Meaning, out goes
stuff like ZoomChangeOverlay and LocationAwareActivity, as well as any
GeoPoint references in Graticule and Info.  Also, updated the Gradle files
so that it builds off of the plain Android 21 SDK, not the APIs one.  That
reference was starting to cause problems with the build.  Guess it's
deprecated now.

The old wiki activities are gone, too.  WikiFragment replaces them all.

Finally, DateButton also goes away, as now I'm using a different popup
and button for the date picker.