Rearranged menu options in ExpeditionMode.
Now, on a phone layout, Graticule and Date will only show up on the action
bar (that is, not in the overflow menu) if an expedition isn't already in
progress.  If it is, those options move out to overflow and Details and
Wiki take their places.  On tablets, however, all four will usually be on
the action bar if an expedition is going on.

So, short story: In general, Details and Wiki will now take precedence on
the menu if there's a valid Info to work with.