Made the WikiFragment header clickable to go to the wiki page.
The header looks like a link now, meaning that, in theory, users will be
drawn to it as something they can tap to launch off to the browser.  I
really couldn't think of a less ugly way to do that, since jamming more
buttons in really didn't look any better.