Converted the legacy branch to use Android Studio.
Make no mistake, I have very little plans of maintaining legacy all that
much beyond a very bare minimum.  It's all new stuff from here on out.
But, since I may want to push a version of GHD that fixes wiki posting
before I'm done with the complete revamp, I may as well make this into
Android Studio, as I'm not really in the mood to reinstall Eclipse to make
this work again.

As I said, this branch won't be as maintained as master, meaning I'm not
really sure I care how well Android Studio is using this data.  As long as
I can get a good APK out of it, it's good enough for me.