Added more null-checking for safety.
It IS possible that mInfoBox or mZoomButtons aren't defined by the time we
get to cleanup in ExpeditionMode.  For instance, if the user doesn't have
the Play Services library installed/updated, has the warning up, then tries
to rotate the device.  Then, everything tries to shut down, cleanup is
called, and we get an NPE.

Good thing I tested this on an AVD that doesn't have the Play Services
installed.  Now, to figure out why said AVD didn't have the Play Services
installed...