Created version history XML file.
Yeah, you heard me.  I'm going crazy.  I'm actually making the version
history data file XML.  I'm going to read it and generate the history list
that way.  And you know what else?  I'm going to make it WELL-FORMED and
VALIDATED XML, too, courtesy of that xsd file in the app/ directory.

I mean, the xsd is for your benefit.  I'm not going to add code to ensure
the schema matches every time the user calls for it.  That's just
ridiculous.  Hence why it isn't in the res/ directories right now.