Removed GHDBasicDialogBuilder.
Yeah, that was a sort of dumb idea.  Now all the dialogs will look like
plain, native Android dialogs.  I'll come up with a different solution
later if I think it's that important.

Note that the wiki login dialog (the thing that pops up when tapping the
notification if a wiki post failed due to a bad login) still uses the GHD
styling, as that can show up anywhere, and I'd like to keep some
identifying mark on it that says it comes from Geohash Droid.