Fixed landscape phone layouts for the date picker.
The "Today" button had a habit of falling off the bottom of the visible
area of the dialog on phones in a landscape orientation.  So, if the screen
is smaller than a tablet (less than sw600dp), that button will now be off
to the right of the date picker widget.