Made version history show up on new versions.
If, on startup, the internal version number is higher than the last one we
reported to the user (or if we've never reported it at all), the version
history dialog will show up.  You have no idea how long that constant was
sitting around in GHDConstants in my local code.