Fixed minor issue with startup not using the right Calendar.
If, on initial startup, the day's stock value isn't posted yet (i.e. the
user is in non-30W and is trying to start before the DJIA opens) and the
user tries to input a date manually, ExpeditionMode wouldn't automatically
change to the appropriate hashpoint (assuming the new date has a valid
stock value).  This change should fix that by ensuring it still goes out
to StockService even if an Info never came back, using the first Location
it grabbed on startup for the Graticule.