Added an initial zoom that doesn't depend on a stock value.
If, for instance, the user started Geohash Droid before the DJIA opens,
the map will remain at its default location (dead zero, the intersection of
the equator and the prime meridian) due to no stock value coming in to tell
it to zoom away.  With this, the initial location will always perform a
zoom to a range of one degree before the stock value comes in.