Initial pass at Marshmallow permissions.
Well, this *works*, per se.  Kind of.  With this in play, all location
requests will pass through a permissions check first and prompt the user
as appropriate.

Unfortunately, it does so a little too aggressively; since this is just
adapting old code, three or four location checks might go up at once (the
empty start, the initial zoom, the InfoBox, the My Location dot on the main
map, etc).  This causes Android to present that popup that many times at
once, and if they all get denied, it gives the error explanation popup that
many times, too.  I'll have to rethink this, but this is where it stands
right now.

Curiously, it seems that MANAGE_DOCUMENTS doesn't require a permissions
check.  I can read images and send them to the wiki without any problems.