Fixed ExpeditionMode not getting the right graticule on an empty start.
Forgot that I can't use mWaitingOnEmptyStart to determine if we need the
empty start Info object, since we're already using that to determine if we
need the empty start location in the first place, and that got cleared once
we got that location.  So, there's a new boolean for the Info portion of
the empty start.