ExpeditionMode ignores DO_INITIAL_START if permissions are denied.
I *think* that's what was doing it.  If it was going to doEmptyStart() from
init() like that, it was completely ignoring any data already in the
bundle, which might happen if the user denies permissions (so the empty
start can't complete, meaning the flag is never cleared) and gets the
graticule manually in SelectAGraticuleMode.