Made CentralMapExtraActivity report location/permissions to fragments.
Since the CentralMapExtraFragments will now need to listen to CentralMap
for location updates, CentralMapExtraActivity will have to do the same job
on a phone (or similar).  This covers the entire lot of changes, both the
fragments and the activities.  But, CentralMap isn't talking to the
fragments yet, so that'll have to come up next.