Moved registering/unregistering the stock receiver to start/stop.
Aha!  I think that finally nails down the bug where it sometimes wouldn't
get a proper info on startup!  The inits on the modes happen during
doReadyChecks (at least when CentralMap initially starts).  doReadyChecks
will do so only after both the map and the API client are ready, and the
last of those to fire is the API client, during onStart.  The init on
ExpeditionMode is when it determines what to do with the bundle it has, be
it do an empty start or load the last known Info.  All nice and good,
except that the stock receiver (needed to listen for stock broadcasts from
StockService and relay them to the modes) was only being registered during
onResume, so there was a chance the API client would finish connecting
beforehand, which in turn meant ExpeditionMode was firing off a stock
lookup before CentralMap was going to be listening for it.

Now, the stock receiver is registered before the API client is asked to
connect, meaning that it'll always be registered before doReadyChecks calls
for the current mode to init.