Updated wiki API location.
Turns out the API location of the wiki changed.  Not sure when it did.
Probably quite a while ago, to be honest.  However, this should be what
it's looking for now.  It also separates out the API URL from what's used
to view pages on the wiki, as those are now under different directories in
the Geohashing wiki.  This still needs proper testing (I'll need to make a
real post on the real wiki), though it might take me longer to get a legacy
build going again...