Put all of StockService's Intent extras into a Bundle.
Turns out there's a bug in some Android firmwares that causes it to not use
the correct ClassLoader when unmarshalling a Parcelable when used over a
remote interface (such as a BroadcastIntent).  This means that things like
Info or Graticule will run into problems when they're sent as such.

StockService sends them as such, as it talks back to everything via
broadcasts.  Thus, it now uses a Bundle to talk back.  Bundles don't try to
unmarshall things until needed, and the ClassLoader can be properly
assigned then.

Thanks to George McBane for helping me track down just what was going on.