Changed the Graticule picker into a widget, not a fragment.
Honestly, I'm not really sure why I made that a fragment to begin with,
and it's probably a much better idea to make it a simpler widget anyway.
However, regardless of anything else, there seems to be some amount of
problem with the picker coming on screen with some devices.  This should
fix that.

I mean, even if it doesn't, I'm keeping it, as this also makes it animate
out of the screen smoother.