Added a way to fade out InfoBox without removing it.
This will be used to make it easier to see the final destination flag or
the current destination marker if either are underneath the area on-screen
where the InfoBox is sitting.  To that end, InfoBox now has a few states
in which it can be, either hidden, faded, or visible.  resolveInfoBoxState
and forceInfoBoxState now handle that.