Made InfoBox fade when it covers the destination or current location.
If the user moves the map such that either the final destination flag or
the current destination marker are directly underneath the InfoBox (and
thus hard to see) now the InfoBox will fade out a bit and become
unclickable until the camera moves it away again.

Note that this doesn't yet account for the view getting resized when, on a
tablet, the extra data fragments (Details and Wiki) appear.  Still not sure
how that'll work out.  In fact, I'm at least a bit sure I'm doing those
very wrong to begin with.