Initial commit of KnownLocationsPicker.
This is how we're going to handle the known locations feature.  We'll have
us a nice good map to handle taps, we'll have us a search field, we'll get
the whole thing wrapped into one activity, and we'll feed that data out to
any other map or whatnot that needs it.  Just, we're not doing that yet.
Still working on it.