More KnownLocations work.
A KnownLocation now takes a range.  Notifications will go up only if the
hashpoint is close enough to a KnownLocation, and "close enough" will be
set on a per-location basis.  They'll be configured by enums, of course,
just to keep it simple (and to prevent users from putting stupid values in,
like something with a greater range than one graticule away).

Also, the MarkerOptions generated from makeMarker won't have any snippet.
That's the responsibility of the caller.