Rearranged some parts of KnownLocationsPicker.
Now the dialog fragment can take a KnownLocation, rather than a boolean, as
a means of keeping track of if the user is editing an existing location or
is inputting a new one.  This goes along with making mMarkerMap into a
BiMap rather than a plain Map so that it can later look up the marker
associated with a given KnownLocation when the dialog is complete (if, for
instance, the screen was rotated, destroying all Markers so it doesn't
know which to edit).

Note that this will most likely be rearranged again once I toss in markers
for search results.