Initial search support for KnownLocationsPicker.
Well, at the very least, you can search for a location.  It won't show up
yet, but at least it'll localize the search to (roughly) the current view
and return whatever the Geocoder thinks should be up to 10 responses.
Trying to search by coordinates won't quite get exact responses, though.
Not sure if I can fix that, to be honest, unless I can come up with some
way to recognize if what was input is some flexible form of coordinates and
override the results.