More KnownLocationsPicker work.
The input box is now on the bottom of the screen.  Completely forgot that
info windows on markers pop up ABOVE the marker, which, combined with the
auto-recenter of that behavior, causes issues.

The EditText also now restricts itself to one line, AND its action button
now does what's expected (that is, conducting the search).

Search results are now drawn on the map (still need a custom marker icon
for those).  They don't do anything yet, nor is there any error reporting.
The addresses will also get redrawn on a restart.

Moving along, the map will only re-center on initial start or after a
search comes back.  MapFragment does a good job of preserving the camera as
need be on restart, so we don't need to do that again.

Still need to handle taps on the addresses and add 'em in.