Changed nearby points to default to on, not off.
At one point, I think it was a bit of a performance hit to keep the eight
nearby points on-map at all times.  Maybe back in Maps API v1.  Nowadays,
though, it's not really a problem, and keeping it on is something that
makes more sense anyway.  So, now it defaults to on.

Also, got rid of GHDConstants.PREFS_BASE.  We don't need that at all
anymore.

And why not, let's just toss another build.gradle update in here.