Added AlarmService functionality for KnownLocations.
Once a stock alarm comes in, AlarmService will handle the entire bit
with checking over KnownLocations and sending out notifications
accordingly.  That part is what's in this commit.  I still need to
test this, though; I'll be doing that soon.

Note that CentralMap isn't set up to handle the new Intent yet, so
that won't work.