Fixed AlarmService.
Well, that's embarrassing.  Remember when I changed StockService to
return all its data in a Bundle in EXTRA_STUFF for weird compatibility
issues?  I forgot to update AlarmService to account for that.  Because
of that, it never became aware that the result had FLAG_ALARM set.
Darn.  That's probably been broken for a while.

Anyway, now that should be fixed.