Encased the AlarmService Info in a Bundle.
Oh, right, that's the problem I had earlier, not all phones can pull out a
Parcelable properly between things like that.  So now the Info that comes
out of AlarmService's notifications will be wrapped back up in a bundle.

Now to figure out if I actually DID code myself into a corner with regards
to making this work in CentralMap...