Made CentralMap and ExpeditionMode handle AlarmService notifications.
Okay, I think I've figured out onNewIntent.  And worked out my now rather
byzantine mess of CentralMapMode code to find a good spot to inject it into
ExpeditionMode.  It ONLY handles it if you're already in CentralMap,
though; it won't handle it if it needs to start it from elsewhere.  Next up
is the onCreate version of this.