Fixed ranges of KnownLocations.
Drat.  Forgot that the distanceTo method returns things in meters, NOT in
kilometers.  The ranges assumed it was in km, meaning it was really
checking if each location was within, say, 2m to 20m, not 2km to 20km.
That WOULD be impressive, but not particularly worth knowing.

To that end, the range settings for KnownLocations are now explicitly in
meters.  Unfortunately, this means that if you were using a pre-release,
you'll need to update all your known locations, else they'll have useless
ranges.  I mean, they had useless ranges already, but now you can actually
give them good ranges.

Aaaaaaaaand updates to Android Studio files.  Again. *sigh*