Replaced the stock HttpClient with a more modern one.
HttpClient, as it stands in Android, is being deprecated.  You can look up
online what exactly happened, but the end result is that it's being
deprecated and that's that.  So now I've switched out all HttpClient stuff
to use a modern Android port of it, one that ISN'T deprecated.  For the
most part, it's identical, but a few minor things changed between the time
the HttpClient 4.0 preview (what Android's stock HttpClient was based off
of) was made by the Apache team to the 4.4 final release.  This should
cover those changes and get rid of the org.apache.http.legacy dependency.
As far as I can tell from testing, this works just as well as the old stuff
did.

As a side effect of this change, the entire local copy of the HttpClient
multipart entity classes is now unnecessary, so out it goes.