Updated KnownLocationPicker's initial zoom to account for ranges.
If the set of KnownLocations in the picker are just right, the auto-zoom
can easily lead to a weird issue where the points themselves are in the
zoom, but the range rings around each point spill out of the view.  So now
the zoom will account for the range, rather than just each point, keeping
all the rings on-screen (under sane conditions).

However, since Android doesn't give me methods to do the inverse geodetic
problem, GHD now includes the OpenSextant Geodesy stuff.  Maven's great,
you know that?