Re-added app.iml.
Oh, for the love of... Android Studio isn't able to auto-generate app.iml
from scratch?  That means that in order for anyone to pull this from the
repo and into Android Studio, this file needs to be here, else they're
going to be forced to manually do a Gradle import to make it work.

Just learned this the hard way when merging master into logcat.  As a part
of the merge, it wiped out app.iml, leaving the logcat branch unbuildable.
Grr...

Looks like I'll just have to keep shoving out updates to this file every
time it decides it wants to shuffle around its XML elements, I guess.