Made sure ExpeditionMode interacts with Last Used Graticule mode better.
If the user has Last Used Graticule mode on but the stock lookup fails
(including if the stock hasn't been posted yet), ExpeditionMode wasn't
reacting properly to a date change.  Now it'll remember what graticule was
given to it (if one has) and use that as the last-chance fallback if the
info hasn't been given yet.