Made StockAlarm start at boot time, preferences permitting.
With the new Known Locations checks and notifications, it makes more sense
to flip StockAlarm on (if need be) at boot time in case the user forgets to
launch the app after a reboot.  Fortunately, it's still kept simple enough
that I don't regret forcing it into the boot process.