Gave Peeron priority over Crox.
Crox appears to be down for the time being (and Peeron isn't), so this will
get things going again.  Note that this is a temporary fix; I'll need to
come up with something more robust to make sure that if Peeron goes down
the same way and Crox comes back up, this won't happen again.