Added hard timeouts to stock fetching and removed StockFetcher.
I suddenly realized I was seriously overcomplicating things with how I was
expecting StockFetcher to work.  All we're really looking for is some way
to do a hard timeout in case a server is giving us problems.  To that end,
now it'll spend at most ten seconds per server before giving up.

I also want some feedback on the main map when a stock fetch is in
progress, so that's next up.