Initial work towards moving the tablet fragments around.
In order to sanely support multiwindow stuff (and because I never really
liked the kluged-together tablet layout I came up with to begin with), the
way the extra fragments (detailed info and wiki posting) work will change
with the Nougat-ready version.  This is the first step of that change: The
fragments now live in the lower-right corner of the map view, rather than
split the screen in half.

This does cause things to crash very very much when resizing the screen in
a way that leaves not enough space for the fragments, but that will be
worked out in time.