Made the extra fragments behave better with multi-window.
In an incredible stroke of good luck, it turns out the way I set up the
layouts is really super good for the way multi-window handles the
activities.  Since it just gives new values to the layout mechanism (for
our purposes, the key is "no longer sw600dp"), we can just check if the
fragment container exists, the same as we already do when deciding if we
want to launch the activity (phones) or put the fragment up (tablets).
This does mean that tablets can now load the activity versions of the extra
fragments if the user attempts to visit them from a multi-window-shrunk
screen, but that's perfectly acceptable.