Made AlarmService a bit more persistent.
So I'm thinking the Known Locations issue has to do with Doze.  It turns
out there's not much I can do about that apart from setting up a Firebase
instance to... um... have it send push messages back to a tiny Android app
for just a once-a-day update of a handful of bytes?  Seriously?  That's the
only way I can assure a network connection during Doze, apart from having
the user whitelist the app?

Regardless, this makes AlarmService be slightly more persistent in getting
a network connection.  Now it'll snooze on a network failure (rather than
just give up) AND I changed it to ignore idle on the alarms (rather than
use the Doze-unfriendly setRepeating method).  Hopefully this makes it try
a bit harder to get the Known Location checks.