Added more strings for tomorrow's and the day after tomorrow's hashpoints.
Before, we only changed the string on the hashpoint markers (both nearby
and active) if the date was explicitly a retro hash.  However, since the
user CAN get tomorrow's and the day after tomorrow's points on a Friday
(Saturday for 30W Rule users), it makes sense to change the strings up when
that's what we're looking at.  Now there's a few new strings.