Added a night mode map.
Now that we have map styles to choose from, I've added an option to set the
map to a nice night mode if need be.  The theme is still bright, though.
I'll try to get that one as soon as I figure out what happened to the
original for the exact version of the GHD background image I use...