More night mode work.
Now there's a BaseGHDThemeActivity.  That handles all the night/day
transition stuff, depending on preferences.  And that preference comes in
via whatever the user set for the map type.

This is still missing a few subtle points, though.  The About window, the
version history, the box for the graticule picker, etc, etc.  Lots of
little things I'll need to work out.