Converted everything over to FusedLocationProviderClient.
The GoogleApiClient method of getting location updates is deprecated,
quite frankly, for good reason.  So now I've switched all of that out
to FusedLocationProviderClient.  That clears up a LOT of odds and ends
I had to do to make GoogleApiClient work, such as a lot of the
permissions checks and such.

However, in retrospect, I had a LOT of weird flags set up to ensure
that GoogleApiClient was connected properly, especially around
doReadyChecks().  Worse, removing the need for the connection callback
revealed that I primarily got lucky in CentralMap when I referenced
mMap before, given there was no guarantee it was defined by that point.
To that end, I've had to poke a lot to clean up problems resulting from
the initial clean up.

Hopefully, this does the trick, but this'll require more testing.  See,
kids, THIS is why you think about testing in the EARLY phases of a
project.  Learn from my experiences!

Also, this clears up a bunch of linty stuff that may have slipped in.
Plus removing the deprecation of onCameraChange in ExpeditionMode.