Added 25 graticule filler option.
This option is specifically for testing the multiple-notifications
feature.  It fills 25 graticules in a 5x5 block centered on Las Vegas
with Known Locations in a similar manner to the other options, only with
a slight random sway to each point.

Also, this adds some sanity checks and cleanup as need be.

This is one of those reasons you don't want to use the logcat branch for
production Geohashing, because it turns out adding 3,025 Known Locations
causes ludicrous slowdown in the Manage Known Locations screen.