Began work on an SQLite-backed version of QueueService.
I'm not entirely certain that keeping that internal queue around in
memory is the most reliable way to deal with this, given how Android
treats Services that aren't doing anything.  To that end, this branch
will be for making an SQLite-backed implementation of it, such that all
Intents are written to the database and read back as a queue as need be.
That should ensure better permanence of the data, and should take away
the weird "one file per Intent, only written if we get a signal to shut
down cleanly" thing I had going before.

At any rate, here's a few abstract methods.  This now breaks WikiService
until I've got the rest of it implemented.