Implemented more methods in QueueService.
Going through what QueueService does on the actual queue end, I needed
a bit more to keep the overall logic the same.  Hence, getQueueCount
and removeNextIntentFromQueue.

On those same lines, getNextIntentFromQueue won't delete the Intent it
returns.  It has to stay there if we got PAUSE back.

Also, the other methods have been renamed so they're not referring to
the database directly.  That seemed a bit implementation-specific.