Removed the Queue parts from QueueService.
Now it's pointing entirely to the database.  I just realized how
inefficient this could get, but it should be a lot more robust and
bulletproof.

Also, this is going to be really fun to feed back into my old test app
I wrote for this seven years ago.  QueueService should probably be
among the first things for which I have proper unit tests written.