Made doKnownLocations more resilient to getClosestInfo failures.
Under normal circumstances, getClosestInfo shouldn't ever return null,
hence why the exception was never caught.  However, in the debug branch,
if the party alarm is manually triggered and makeNineThirty thinks it
should be "tomorrow", that'll cause getClosestInfo to check the database
for Infos on dates that don't have stocks yet, crashing the app.  So now
it'll just skip over anything that throws an exception under the
assumption that you're debugging at that point.