Added a ready check to onResume in CentralMap.
Good heavens, I wound up with some convoluted code to make sure the map
fragment was ready.  Regardless, this should ensure the known location
markers are redrawn when re-entering CentralMap from Preferences.  That
is, if the user changes the markers or turns them all on/off, now it
should be reflected properly.