Made top toolbar titles work again.
Since the toolbars don't tie into the Activity titles, all calls that
would set the titles that way will need to go through the toolbars now.
As an added bonus, ExpeditionMode now makes use of the toolbar's
subtitle for the date, meaning I've got space to make it clear that the
user's doing a retrohash, if appropriate.

I may need to tweak the logo for spacing purposes.