Changed the meaning of "pause" in QueueService.
Now, any time the queue isn't servicing Intents, it'll stop entirely.
This differs from previous versions where a "paused" queue would try to
remain running, per se, expecting to be able to write out all pending
Intents at onDestroy time.  Turns out that isn't at all reliable.  This,
on the other hand, should be; it does whatever task it needs to, and if
it pauses, it writes the queue out and stops, expecting to always reload
the queue from storage on start.

This removes the mIsPaused boolean and anything associated with it, is
what I'm saying.