Made PreferencesScreen into something androidx-y.
This SHOULD make it act like a more modern Android preference screen.
It doesn't, though; it crashes because the theme isn't part of the
Theme.AppCompat hierarchy.  And it's that way because I'm doing goofy
things to fake night mode throughout the app.

So the end result is that I may need to jump back to updating things to
use the PROPER DayNight theme (which IS part of Theme.AppCompat) and
thus rework the UI around that.