Changed WikiException to take an error code, not a text ID.
It turns out that in Gradle 5.0, resource IDs won't necessarily be final
(and thus consistent), so this is step one of resolving that.  Now,
WikiException directly takes an error code from the wiki, unless it's
not a wiki-spawned error code, in which case it takes some internal
code.  This is probably the wrong way to do this.