Moved zoom button to the lower-right.
Okay, I've got an idea for the toolbar issue.  The short version: Remove
the lower toolbar and move all those options into another image button
like the zoom controls.  First step, right-align stuff.  The zoom
controls are now on the lower-right, not the lower-left.  This may
differ in RtL languages.