Rearranged ZoomButtons to group things and be more horizontal.
Here's the plan: ZoomButtons will be refactored to contain more than
just the zoom buttons.  It'll also contain the whole group of wiki,
date, details, and graticule picker.  It'll also be renamed.  That'll
get rid of the lower toolbar.

As such, the zoom buttons are now horizontal.  They'll still fly in from
the side.  The top-level buttons will be vertical.  More to come.