Made it so images are resolved before they get to WikiService.
It turns out my clever idea to just pass around image URIs to save
memory only works if the Context that opens said URIs has permission to
do so.  That permission is only granted to whatever Context sent out the
ACTION_GET_CONTENT/ACTION_OPEN_DOCUMENT Intents.  The end result is that
WikiService couldn't reliably open image URIs, meaning all wiki image
posting was broken.

With this change, all image opening is handled in WikiFragment,
resulting in the byte[] being created there and stuffed into the Intent
that WikiService eventually reads, meaning all WikiService has to do is
send that on its merry way.  It's serialized out with Base64, so the
database won't have an issue with it. (#77)