Converted JobServices into WorkMananger things.
Both AlarmService and WikiService were using JobService to react to the
internet connection coming back.  While this worked, more or less, it
wasn't available in Android versions prior to 21.  The pre-21 system
used manifest-declared services and receivers to handle that sort of
thing, which was ugly and possibly a drain on the battery regardless.

Thus, this change switches to using AndroidX's WorkManager.  That's
compatible with both, collapses things down a lot, removes the services
and receivers from the manifest, and simplifies stuff greatly. (#66)