Changed WikiImageUtils to read EXIF data for ImageInfo.
This is more or less how this used to work, back before I switched
things out when Android changed how document opening worked for images.
This means that ImageInfo will now have more accurate data, which in
turn should mean, for instance, a user can post an image sometime after
they took it and have the location and time show up as what's in the
image, not where/when they are when the post is made.