Reverted all the recent ZoomButtons changes.
On second thought, I don't think that's the right way to go about this.
MenuButtons is now ZoomButtons again, the zoom controls are vertically-
oriented, and the button's back to the lower-left.  I'll think about
what to do with it later.  I've got other ideas for the toolbar, if need
be.