Made ImageInfo's constructor ignore the timestamp param for now.
Apparently what I use to get the timestamp from an image isn't behaving
the way I want it to.  Since the only reason I use that timestamp is to
generate unique filenames for wiki uploads, for the time being it'll
just use the time at the moment the ImageInfo object was created. (#80)