Added a whole bunch of image rotatey and flippy stuff.
All (or most) of this will come in really handy when it's time to apply
rotation and flipping and stuff.  It'll make a lot more sense once
there's buttons on the wiki post screen that allows for rotation and
flipping before posting.

Now, to figure out how to fit editing buttons on the wiki post screen...