Added a bunch of stuff regarding the image editing UI.
I realized all of this ought to be in a different branch, so now it is.
This also adds in the editing UI images (such as they are for now),
some layout (incomplete; still need the tablet-sized versions), a new
menu, and some interaction in WikiFragment.  It's absolutely not
finished yet.  I think I'll need to redo more of the wiki UI in order
to make it fit, come to think of it.