Made PreferencesScreen only load the main fragment if the stack's empty.
With the day/night switch now restarting the activity on a config
change, there's a way that PreferencesScreen can come into onCreate with
a populated fragment stack.  That, in turn, means the interface will
take one extra back button press to get out of preferences for every
time the user switched between light and dark modes, as nothing either
emptied the stack or checked if there was a stack there before replacing
the top with MainPreferenceFragment.

The short version is that if the stack's got something in it, it won't
do the replace, so not only will there not be extra stack entries lying
around, it'll go back to the correct page of the preferences screen on a
config change.