Added permission checks around all notifications.
This one's troublesome.  There were a bunch of linting errors indicating
that notifications require permissions which can be denied by the user.
While this is true, and I believe has been true for a while now, this
does remind me that if the user has turned off all notifications or they
default to off, they might not see a lot of important things (i.e. bad
wiki password).  I may need to review that later.