Added Type to Somethingicule and updated Graticule serialization.
My plans for Centicule now revolve around just making it handle ints
from 0 to 899 or 1799, which means they'll serialize to the same fields
that Graticule uses.  That, in turn, means we need an enum to determine
which type of Somethingicule we're deserializing.

THAT, in turn, means that the Somethingicule deserializer can just call
deserializeFromJSON() for the entire JSONObject, which simplifies its
code significantly.